refactor: extract terminal-bench environment cli commands
