+++ loopx/cli_commands/benchmark_review_lifecycle.py
@@ -207,243 +207,6 @@ def render_benchmark_lifecycle_state_markdown(payload: dict[str, object]) -> str
     ) + "\n"
 
 
-def render_terminal_bench_environment_setup_gate_markdown(
-    payload: dict[str, object],
-) -> str:
-    capability = (
-        payload.get("harbor_run_help_capability")
-        if isinstance(payload.get("harbor_run_help_capability"), dict)
-        else {}
-    )
-    contract = (
-        payload.get("probe_contract")
-        if isinstance(payload.get("probe_contract"), dict)
-        else {}
-    )
-    read_boundary = (
-        payload.get("read_boundary")
-        if isinstance(payload.get("read_boundary"), dict)
-        else {}
-    )
-    lines = [
-        "# Terminal-Bench Environment Setup Gate",
-        "",
-        f"- Schema: `{payload.get('schema_version')}`",
-        f"- Benchmark: `{payload.get('benchmark_id')}`",
-        f"- Task: `{payload.get('task_id')}`",
-        f"- Preflight ready: `{payload.get('preflight_ready')}`",
-        "- Previous setup failure: "
-        f"`{payload.get('previous_environment_setup_failure_present')}`",
-        f"- Help probe ok: `{capability.get('probe_ok')}`",
-        f"- Direct setup-only route: `{payload.get('direct_setup_only_route_allowed')}`",
-        "- NOP disable-verification route: "
-        f"`{payload.get('nop_disable_verification_probe_allowed')}`",
-        "- Environment setup probe allowed: "
-        f"`{payload.get('environment_setup_probe_allowed')}`",
-        f"- Same-task repeat allowed: `{payload.get('same_task_repeat_allowed')}`",
-        f"- First blocker: `{payload.get('first_blocker')}`",
-        f"- Next action: {payload.get('next_allowed_action')}",
-        f"- Probe agent: `{contract.get('agent')}`",
-        f"- No upload / submit eligible: `{contract.get('no_upload')}` / `{contract.get('submit_eligible')}`",
-        f"- Codex invoked: `{contract.get('codex_invoked')}`",
-        f"- Compact only: `{read_boundary.get('compact_only')}`",
-        f"- Raw help recorded: `{read_boundary.get('raw_help_recorded')}`",
-    ]
-    return "\n".join(lines) + "\n"
-
-
-def render_terminal_bench_environment_setup_probe_launch_markdown(
-    payload: dict[str, object],
-) -> str:
-    post_launch = (
-        payload.get("post_launch_materialization")
-        if isinstance(payload.get("post_launch_materialization"), dict)
-        else {}
-    )
-    boundary = payload.get("boundary") if isinstance(payload.get("boundary"), dict) else {}
-    lines = [
-        "# Terminal-Bench Environment Setup Probe Launch",
-        "",
-        f"- Schema: `{payload.get('schema_version')}`",
-        f"- Dry run: `{payload.get('dry_run')}`",
-        f"- Run: `{payload.get('run_basename')}`",
-        f"- Job: `{payload.get('job_name')}`",
-        f"- Process started: `{payload.get('process_started')}`",
-        f"- Process state: `{payload.get('process_state')}`",
-        f"- Return code: `{payload.get('returncode')}`",
-        f"- Timed out: `{payload.get('process_timed_out')}`",
-        f"- Materialization wait seconds: `{payload.get('materialization_wait_seconds')}`",
-        f"- Materialization wait timed out: `{payload.get('materialization_wait_timed_out')}`",
-        f"- First blocker: `{payload.get('first_blocker')}`",
-        f"- Compact failure: `{payload.get('compact_failure_class')}`",
-        f"- Ready for launch state: `{payload.get('ready_for_launch_state')}`",
-        f"- Ready for compact ingest: `{payload.get('ready_for_compact_result_ingest')}`",
-        f"- Ready for failure marker: `{payload.get('ready_for_compact_failure_marker')}`",
-        f"- Post-launch blocker: `{post_launch.get('first_blocker')}`",
-        f"- No upload / submit eligible: `{boundary.get('no_upload')}` / `{boundary.get('submit_eligible')}`",
-        f"- Raw logs read: `{boundary.get('raw_logs_read')}`",
-        f"- Task text read: `{boundary.get('task_text_read')}`",
-    ]
-    return "\n".join(lines) + "\n"
-
-
-def render_terminal_bench_worker_materialization_probe_launch_markdown(
-    payload: dict[str, object],
-) -> str:
-    post_launch = (
-        payload.get("post_launch_materialization")
-        if isinstance(payload.get("post_launch_materialization"), dict)
-        else {}
-    )
-    boundary = payload.get("boundary") if isinstance(payload.get("boundary"), dict) else {}
-    command_shape = (
-        payload.get("command_shape")
-        if isinstance(payload.get("command_shape"), dict)
-        else {}
-    )
-    lines = [
-        "# Terminal-Bench Worker Materialization Probe Launch",
-        "",
-        f"- Schema: `{payload.get('schema_version')}`",
-        f"- Dry run: `{payload.get('dry_run')}`",
-        f"- Run: `{payload.get('run_basename')}`",
-        f"- Job: `{payload.get('job_name')}`",
-        f"- Process started: `{payload.get('process_started')}`",
-        f"- Process state: `{payload.get('process_state')}`",
-        f"- Return code: `{payload.get('returncode')}`",
-        f"- Timed out: `{payload.get('process_timed_out')}`",
-        f"- Resume after materialization: `{payload.get('resume_after_materialization')}`",
-        f"- Resume attempted: `{payload.get('resume_after_materialization_attempted')}`",
-        f"- First blocker: `{payload.get('first_blocker')}`",
-        f"- Compact failure: `{payload.get('compact_failure_class')}`",
-        f"- Ready for launch state: `{payload.get('ready_for_launch_state')}`",
-        f"- Ready for compact ingest: `{payload.get('ready_for_compact_result_ingest')}`",
-        f"- Post-launch blocker: `{post_launch.get('first_blocker')}`",
-        "- Probe-only kwarg: "
-        f"`{command_shape.get('worker_materialization_probe_only')}`",
-        f"- No upload / submit eligible: `{boundary.get('no_upload')}` / `{boundary.get('submit_eligible')}`",
-        f"- Task solver invoked by probe: `{boundary.get('task_solver_invoked_by_probe')}`",
-        f"- Raw logs read: `{boundary.get('raw_logs_read')}`",
-        f"- Task text read: `{boundary.get('task_text_read')}`",
-    ]
-    return "\n".join(lines) + "\n"
-
-
-def render_terminal_bench_case_run_launch_markdown(
-    payload: dict[str, object],
-) -> str:
-    post_launch = (
-        payload.get("post_launch_materialization")
-        if isinstance(payload.get("post_launch_materialization"), dict)
-        else {}
-    )
-    boundary = payload.get("boundary") if isinstance(payload.get("boundary"), dict) else {}
-    command_shape = (
-        payload.get("command_shape")
-        if isinstance(payload.get("command_shape"), dict)
-        else {}
-    )
-    lines = [
-        "# Terminal-Bench Case Run Launch",
-        "",
-        f"- Schema: `{payload.get('schema_version')}`",
-        f"- Dry run: `{payload.get('dry_run')}`",
-        f"- Run: `{payload.get('run_basename')}`",
-        f"- Job: `{payload.get('job_name')}`",
-        f"- Process started: `{payload.get('process_started')}`",
-        f"- Process state: `{payload.get('process_state')}`",
-        f"- Return code: `{payload.get('returncode')}`",
-        f"- Timed out: `{payload.get('process_timed_out')}`",
-        f"- First blocker: `{payload.get('first_blocker')}`",
-        f"- Compact failure: `{payload.get('compact_failure_class')}`",
-        f"- Ready for launch state: `{payload.get('ready_for_launch_state')}`",
-        f"- Ready for compact ingest: `{payload.get('ready_for_compact_result_ingest')}`",
-        f"- Post-launch blocker: `{post_launch.get('first_blocker')}`",
-        "- Probe-only kwarg: "
-        f"`{command_shape.get('worker_materialization_probe_only')}`",
-        f"- No upload / submit eligible: `{boundary.get('no_upload')}` / `{boundary.get('submit_eligible')}`",
-        f"- Task solver invoked: `{boundary.get('task_solver_invoked')}`",
-        f"- Model API expected: `{boundary.get('model_api_expected')}`",
-        f"- Raw logs read: `{boundary.get('raw_logs_read')}`",
-        f"- Task text read: `{boundary.get('task_text_read')}`",
-    ]
-    return "\n".join(lines) + "\n"
-
-
-def render_terminal_bench_worker_materialization_probe_poll_markdown(
-    payload: dict[str, object],
-) -> str:
-    post_launch = (
-        payload.get("post_launch_materialization")
-        if isinstance(payload.get("post_launch_materialization"), dict)
-        else {}
-    )
-    boundary = payload.get("boundary") if isinstance(payload.get("boundary"), dict) else {}
-    pid_state = (
-        payload.get("pid_state") if isinstance(payload.get("pid_state"), dict) else {}
-    )
-    lines = [
-        "# Terminal-Bench Worker Materialization Probe Poll",
-        "",
-        f"- Schema: `{payload.get('schema_version')}`",
-        f"- Run: `{payload.get('run_basename')}`",
-        f"- Job: `{payload.get('job_name')}`",
-        f"- Process state: `{payload.get('process_state')}`",
-        f"- PID file present/parsed: `{pid_state.get('pid_file_present')}`/`{pid_state.get('pid_parse_ok')}`",
-        f"- First blocker: `{payload.get('first_blocker')}`",
-        f"- Compact failure: `{payload.get('compact_failure_class')}`",
-        f"- Ready for launch state: `{payload.get('ready_for_launch_state')}`",
-        f"- Ready for compact ingest: `{payload.get('ready_for_compact_result_ingest')}`",
-        f"- Ready for failure marker: `{payload.get('ready_for_compact_failure_marker')}`",
-        f"- Post-launch blocker: `{post_launch.get('first_blocker')}`",
-        f"- No upload / submit eligible: `{boundary.get('no_upload')}` / `{boundary.get('submit_eligible')}`",
-        f"- Raw logs read: `{boundary.get('raw_logs_read')}`",
-        f"- Task text read: `{boundary.get('task_text_read')}`",
-        f"- Command line read: `{boundary.get('command_line_read')}`",
-    ]
-    return "\n".join(lines) + "\n"
-
-
-def render_terminal_bench_resume_observation_markdown(
-    payload: dict[str, object],
-) -> str:
-    post_launch = (
-        payload.get("post_launch_materialization")
-        if isinstance(payload.get("post_launch_materialization"), dict)
-        else {}
-    )
-    boundary = payload.get("boundary") if isinstance(payload.get("boundary"), dict) else {}
-    command_shape = (
-        payload.get("command_shape")
-        if isinstance(payload.get("command_shape"), dict)
-        else {}
-    )
-    lines = [
-        "# Terminal-Bench Job Resume",
-        "",
-        f"- Schema: `{payload.get('schema_version')}`",
-        f"- Dry run: `{payload.get('dry_run')}`",
-        f"- Run: `{payload.get('run_basename')}`",
-        f"- Job: `{payload.get('job_name')}`",
-        f"- Process started: `{payload.get('process_started')}`",
-        f"- Process state: `{payload.get('process_state')}`",
-        f"- Return code: `{payload.get('returncode')}`",
-        f"- Timed out: `{payload.get('process_timed_out')}`",
-        f"- First blocker: `{payload.get('first_blocker')}`",
-        f"- Compact failure: `{payload.get('compact_failure_class')}`",
-        f"- Ready for launch state: `{payload.get('ready_for_launch_state')}`",
-        f"- Ready for compact ingest: `{payload.get('ready_for_compact_result_ingest')}`",
-        f"- Ready for failure marker: `{payload.get('ready_for_compact_failure_marker')}`",
-        f"- Post-launch blocker: `{post_launch.get('first_blocker')}`",
-        f"- Uses Harbor job resume: `{command_shape.get('uses_harbor_job_resume')}`",
-        f"- No upload / submit eligible: `{boundary.get('no_upload')}` / `{boundary.get('submit_eligible')}`",
-        f"- Resume invoked: `{boundary.get('resume_invoked')}`",
-        f"- Raw logs read: `{boundary.get('raw_logs_read')}`",
-        f"- Task text read: `{boundary.get('task_text_read')}`",
-    ]
-    return "\n".join(lines) + "\n"
-
-
 def render_benchmark_verifier_attribution_review_markdown(
     payload: dict[str, object],
 ) -> str:
